Fix `@FILE` arguments.
When `@FILE` arguments are passed, read the whole file not just the
first line like the current implementation.

Drop the requirement that `@...` arguments have to end in `.rsp`. This
is consistent with dmd and used out in the wild by dub which uses
`@...rsp` when building and `@...lnk` when linking. To account for
this the matching code has been moved up the if .. elsif block.

The implementation was also changed to use a local file handle instead
of the legacy bareword and the two-argument form of open has been
replaced by the three-argument one.

Add a space in the help string to compensate for the `\@` escape.

When a string argument ends in a number of `\`, half their number
instead of dropping it to one, this is the same as dmd. It is unlikely
that this will ever come up so it is only for
completeness. Specifically `"a \\\\"` gets converted to `a \` when dmd
converts it to `a \\`.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>