Fix docs about -version=level and -debug=level
Although removed in gdc-13, -version and -debug with a integer value
are still supported by gdc-12. The documentation, both in this
package's help text and man page, as well as in the gdc man page, as
well as in the (old) upstream dmd man page say that the -version=level
switch is supposed to compile code in blocks >= level, but a simple
compile run gives contrary results, those being blocks <= level are
compiled. This feature has been removed so no point it thinking too
much about it so have the docs document the observed behavior.

Also fix the man page having an incomplete -version=ident entry.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>