dmd-script: Expand @rsp arguments in-place in ARGV
When expanding response files place the extracted arguments in the
same spot that the response argument appeared on, in the ARGV array,
instead of appending them at the end.

Specifically make `gdmd @args -foo` expand to `gdmd -arg1 -arg2 -foo`,
not `gdmd -foo -arg1 -arg2`.

This is especially problematic with then `-run` flag since `gdmd @args
-run file` will map to `gdmd -run file -arg1 -arg2` which no longer
passes `-arg1 -arg2` to gdc, but to the compiled program.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>